Create inverse Polish expression
Application of stack: (inverse Polish expression) input the value of a string expression, assuming that there are only plus (+), minus (-), multiply (*), divide (/) in the string, brackets () can be nested, and the operands are positive integers. Output the inverse Polish expression of the string, and calculate the value of the string expression. For example, when the input string is 5 + (12-4 * 2) * 6-9, the output: inverse Polish expression is 5, 12, 4, 2, *, -, 6, *, +, 9, -. Its output value is 20.