Create Pointer, string and linked list
Pointer, string and linked list: input n strings from the keyboard, create a single linked list in increasing order according to the dictionary, and output it according to the dictionary order. The single linked list is inverted and output in the reverse order of dictionary order. Then enter an M, where (n > = m), delete the m-th largest string from the list, and output in reverse order. (Note: the dynamic pointer is used to allocate the space for the string of the linked list node, and the maximum length of the input character is not more than 1000)